Move all paths from constants to constants/paths
