+++ src/backend/config.ts
@@ -5,15 +5,7 @@ import {
   GlobalConfigVersion,
   WineInstallation
 } from 'common/types'
-import {
-  currentGlobalConfigVersion,
-  configPath,
-  defaultWinePrefix,
-  gamesConfigPath,
-  heroicInstallPath,
-  userHome,
-  getSteamCompatFolder
-} from './constants'
+import { currentGlobalConfigVersion, getSteamCompatFolder } from './constants'
 
 import { logError, logInfo, LogPrefix } from './logger/logger'
 import {
@@ -35,6 +27,13 @@ import {
   isIntelMac,
   isWindows
 } from './constants/environment'
+import {
+  configPath,
+  defaultWinePrefix,
+  gamesConfigPath,
+  heroicInstallPath,
+  userHome
+} from './constants/paths'
 
 /**
  * This class does config handling.
