+++ src/backend/constants/paths.ts
@@ -1,7 +1,10 @@
 import { app } from 'electron'
 import { mkdirSync } from 'graceful-fs'
-import { join } from 'path'
+import { homedir } from 'os'
+import { join, resolve } from 'path'
+import { env } from 'process'
 import { dirSync } from 'tmp'
+import { isSnap } from './environment'
 
 let configFolder = app.getPath('appData')
 // If we're running tests, we want a config folder independent of the normal
@@ -15,9 +18,30 @@ if (process.env.CI === 'e2e') {
   mkdirSync(join(configFolder, 'heroic'))
 }
 
+export const flatpakHome = env.XDG_DATA_HOME?.replace('/data', '') || homedir()
+export const userHome = isSnap ? env.SNAP_REAL_HOME! : homedir()
+
 export const appFolder = join(configFolder, 'heroic')
 export const userDataPath = app.getPath('userData')
 export const toolsPath = join(appFolder, 'tools')
 export const epicRedistPath = join(toolsPath, 'redist', 'legendary')
 export const runtimePath = join(toolsPath, 'runtimes')
 export const defaultUmuPath = join(runtimePath, 'umu', 'umu_run.py')
+export const configPath = join(appFolder, 'config.json')
+export const gamesConfigPath = join(appFolder, 'GamesConfig')
+export const heroicIconFolder = join(appFolder, 'icons')
+export const heroicInstallPath = join(userHome, 'Games', 'Heroic')
+export const defaultWinePrefixDir = join(
+  userHome,
+  'Games',
+  'Heroic',
+  'Prefixes'
+)
+export const defaultWinePrefix = join(defaultWinePrefixDir, 'default')
+export const fixesPath = join(appFolder, 'fixes')
+
+export const publicDir = resolve(
+  __dirname,
+  '..',
+  app.isPackaged || process.env.CI === 'e2e' ? '' : '../public'
+)
