+++ src/backend/storeManagers/legendary/library.ts
@@ -25,7 +25,7 @@ import {
   getFileSize,
   axiosClient
 } from '../../utils'
-import { fallBackImage, legendaryLogFile, userHome } from '../../constants'
+import { fallBackImage, legendaryLogFile } from '../../constants'
 import {
   logDebug,
   logError,
@@ -53,6 +53,7 @@ import { Entries } from 'type-fest'
 import { runLegendaryCommandStub } from './e2eMock'
 import { legendaryConfigPath, legendaryMetadata } from './constants'
 import { isLinux, isWindows } from 'backend/constants/environment'
+import { userHome } from 'backend/constants/paths'
 
 const allGames: Set<string> = new Set()
 let installedGames: Map<string, InstalledJsonMetadata> = new Map()
