+++ src/backend/tray_icon/tray_icon.ts
@@ -6,10 +6,11 @@ import { handleProtocol } from '../protocol'
 import { getRecentGames, maxRecentGames } from '../recent_games/recent_games'
 import { handleExit, showAboutWindow } from '../utils'
 import { GlobalConfig } from '../config'
-import { fixAsarPath, publicDir } from '../constants'
+import { fixAsarPath } from '../constants'
 import { backendEvents } from '../backend_events'
 import { join } from 'node:path'
 import { isMac } from 'backend/constants/environment'
+import { publicDir } from 'backend/constants/paths'
 
 const iconDark = fixAsarPath(join(publicDir, 'icon-dark.png'))
 const iconLight = fixAsarPath(join(publicDir, 'icon-light.png'))
